+++ src/components/Icon/index.ts
@@ -1,0 +1,2 @@
+export * from './Icon';
+export * from './IconSet';
